Restructure the advisory harness into a per-advisory probe package
Replaces the single 859-line audit/advisories.py (plus a committed ~90 KB
advisories.json) with nltk/test/unit/security_probes/: one ghsa_<id>.py module
per advisory, each registering a probe that runs the attack the advisory
describes and reports what the tree does now.

Three problems with the previous shape drove this, each a real exposure in a
security PR:

- The coverage test reached outside the test tree with
  `pathlib.Path(__file__).resolve().parents[3] / "audit"` and then
  `sys.path.insert(0, ...)` + `import advisories`. In an installed package
  parents[3] is site-packages, so the test would inject site-packages/audit
  onto sys.path[0] and import whatever `advisories` module is first there --
  arbitrary code execution by path confusion, the very class these advisories
  are about. The package now lives inside nltk/test/unit and is imported as a
  normal sibling; nothing computes an outside path or touches sys.path. The
  STATIC probes locate NLTK source via each module's own __file__, not path
  arithmetic, for the same reason.

- The committed advisories.json was large and was loaded at test time. An
  advisory description is text written by a reporter; parsing an ever-growing
  untrusted document during the test is its own exposure. No JSON is committed
  now. The offline tests need none, and the live list is pulled from GitHub in
  one isolated CI job.

- One 859-line file is not reviewable per advisory. Each advisory is now its
  own small, independently readable module.

Two test entry points:

- test_advisory_probes.py (offline, runs on every pytest): runs every probe and
  fails on any VULNERABLE, and includes negative controls that break each guard
  in-process and assert the probe flips to VULNERABLE -- so a probe cannot pass
  by never exercising the attack. 8 tests, ~3s, no network.

- test_advisory_coverage_ci.py (CI only): pulls the published advisory list from
  GitHub and fails if a newly published advisory has no probe. Gated behind
  NLTK_ADVISORY_CI, which is set in exactly one place -- a dedicated
  advisory-coverage job with contents:read -- so a normal pytest run never
  touches the network. It skips, never fails, if the fetch itself cannot be
  made.

Result on develop: 38 probes, 31 FIXED, 7 STATIC, 0 VULNERABLE. STATIC means
the guard is confirmed in source but the attack was not executed (needs a
network peer, a real download or a running JVM); it is labelled so a source
inspection is never passed off as an exploit attempt.

ci.yml: add an offline "Run advisory-verification probes" step alongside the
existing isolated security steps, and a separate advisory-coverage job for the
GitHub gap check.
