+++ nltk/test/unit/security_probes/__main__.py
@@ -1,0 +1,7 @@
+"""Print the advisory probe report: ``python -m nltk.test.unit.security_probes``."""
+
+import sys
+
+from . import run
+
+sys.exit(1 if run() else 0)
