+++ nltk/test/unit/security_probes/ghsa_5wp5_5229_5g6q.py
@@ -1,0 +1,16 @@
+"""GHSA-5wp5-5229-5g6q [high] -- Missing Post-Download Integrity Verification Allows Malicious Package Injection
+
+Probe: run the attack, report what the tree does now. See _base for statuses.
+"""
+from ._base import STATIC, VULNERABLE, probe, read_source
+
+
+@probe("GHSA-5wp5-5229-5g6q")
+def _downloader_integrity():
+    """No integrity verification between download and extraction."""
+    source = read_source("nltk.downloader")
+    markers = ("checksum", "sha256", "hashlib", "digest")
+    hits = [m for m in markers if m in source]
+    if not hits:
+        return VULNERABLE, "downloader.py performs no post-download verification"
+    return STATIC, "integrity check present (%s)" % ", ".join(hits[:3])
