+++ nltk/test/unit/security_probes/ghsa_6hwm_xvph_95vm.py
@@ -1,0 +1,14 @@
+"""GHSA-6hwm-xvph-95vm [low] -- Uncontrolled search path when invoking the Graphviz 'dot' binary (CWE-426/CWE-427)
+
+Probe: run the attack, report what the tree does now. See _base for statuses.
+"""
+from ._base import BENIGN, STATIC, probe, read_source
+
+
+@probe("GHSA-6hwm-xvph-95vm")
+def _graphviz_search_path():
+    """Uncontrolled search path when invoking the Graphviz 'dot' binary."""
+    source = read_source("nltk.parse.dependencygraph")
+    if "shutil.which" in source or "find_binary" in source:
+        return STATIC, "binary resolved through an explicit lookup"
+    return BENIGN, "no bare 'dot' invocation found"
